src/requirements/test.txt
factory-boy
flake8
isort
pytest
pytest-cov
pytest-django
pytest-echo
pytest-pythonpath
